+++ tests/testthat/test-utilities-scenario-configuration.R
@@ -86,11 +86,40 @@ test_that("It creates multiple correct scenarios", {
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$steadyStateTime, 500)
 })
 
+
+test_that("It creates a population scenario", {
+  scenarioNames <- "PopulationScenario"
+
+  scenarioConfigurations <- readScenarioConfigurationFromExcel(
+    scenarioNames = scenarioNames,
+    projectConfiguration = projectConfiguration
+  )
+  expect_equal(names(scenarioConfigurations), scenarioNames)
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$applicationProtocol, "Aciclovir_iv_250mg")
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$individualId, "Indiv")
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$populationId, "TestPopulation")
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$modelFile, "Aciclovir.pkml")
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$paramSheets, enum(enumValues = "Global"))
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$pointsPerMinute, 1)
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$scenarioName, "PopulationScenario")
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$setTestParameters, FALSE)
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulateSteadyState, FALSE)
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulationTime, toUnit(
+    quantityOrDimension = ospDimensions$Time,
+    values = 12,
+    targetUnit = ospUnits$Time$min,
+    sourceUnit = ospUnits$Time$h
+  ))
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulationRunOptions, NULL)
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulationType, "Population")
+  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$steadyStateTime, 1000)
+})
+
 test_that("It creates all scenarios if no name is defined", {
   scenarioConfigurations <- readScenarioConfigurationFromExcel(
     projectConfiguration = projectConfiguration
   )
-  expect_equal(names(scenarioConfigurations), c("TestScenario", "TestScenario2"))
+  expect_equal(names(scenarioConfigurations), c("TestScenario", "TestScenario2", "PopulationScenario"))
 })
 
 test_that("It correctly applies a custom function", {
@@ -134,7 +163,6 @@ test_that("It throws an error when trying to set wrong arguments for the custom 
     projectConfiguration = projectConfiguration
   )
 
-
   customFunction <- function(doseFactor) {
     doseParamPath <- "Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose"
     doseParam <- getParameter(
@@ -147,4 +175,18 @@ test_that("It throws an error when trying to set wrong arguments for the custom 
   scenarioConfigurations$TestScenario$customFunction <- customFunction
 
   expect_error(scenarioConfigurations$TestScenario$customFunctionArgs <- list(wrong = 2), messages$errorWrongArguments("doseFactor"))
+})
+
+test_that("It throws an error when no population is specified for a population scenario", {
+  scenarioNames <- c("TestScenario",
+                     "PopulationScenario")
+
+  scenarioConfigurations <- readScenarioConfigurationFromExcel(
+    scenarioNames = scenarioNames,
+    projectConfiguration = projectConfiguration
+  )
+
+  scenarioConfigurations$PopulationScenario$populationId <- NULL
+
+  expect_error(runScenarios(scenarioConfigurations), messages$noPopulationIdForPopulationScenario("PopulationScenario"))
 })